This is my last commit for the TextChat component. You have all the information about usage in the README.md file and in the docs folder.
At functionality level, I would want to add a preprocess method on the
widget to allow introducing commands or trimming the message. Actually
nothing impossible right now, but I would want to ship the Widget with
that as well.

At CSS level, @tfish has the visuals. I would want to adjust the sizes
to match the specs when the .ot-texchat font-size is 12pt but expressing
all in ems to improve responsiveness.

The testing framework is based on require.js and Karma to support
multiple browsers. Almost all the classes are tested but ChatUI has been
only drafted and need a lot of tests to be implemented. My
recommendation is to include jquery to ease the simulation of events of
some specific library to simulate front-end interactions.